Split out bulkServices from services
