+++ src/modules/services/components/EnrollmentServicesPage.tsx
@@ -1,71 +1,23 @@
 import AddIcon from '@mui/icons-material/Add';
-import { Button, Typography } from '@mui/material';
-import { Stack } from '@mui/system';
+import { Button } from '@mui/material';
 import { useState } from 'react';
 
-import { ColumnDef } from '@/components/elements/table/types';
 import TitleCard from '@/components/elements/TitleCard';
 import NotFound from '@/components/pages/NotFound';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import useEnrollmentDashboardContext from '@/modules/enrollment/hooks/useEnrollmentDashboardContext';
 import { useFilters } from '@/modules/hmis/filterUtil';
-import { parseAndFormatDate, serviceDetails } from '@/modules/hmis/hmisUtil';
 import { useServiceDialog } from '@/modules/services/hooks/useServiceDialog';
+import {
+  SERVICE_BASIC_COLUMNS,
+  SERVICE_COLUMNS,
+} from '@/modules/services/serviceColumns';
 import {
   GetEnrollmentServicesDocument,
   GetEnrollmentServicesQuery,
   GetEnrollmentServicesQueryVariables,
-  ServiceBasicFieldsFragment,
   ServiceFieldsFragment,
 } from '@/types/gqlTypes';
-
-export const SERVICE_BASIC_COLUMNS: {
-  [key: string]: ColumnDef<ServiceBasicFieldsFragment>;
-} = {
-  dateProvided: {
-    header: 'Date Provided',
-    linkTreatment: true,
-    render: (s) => parseAndFormatDate(s.dateProvided),
-  },
-  serviceType: {
-    header: 'Service Type',
-    render: ({ serviceType }) => {
-      if (!serviceType) return 'Unknown Service';
-      const { name, category } = serviceType;
-      if (name === category) return name;
-      return `${category} - ${name}`;
-    },
-  },
-};
-
-export const SERVICE_COLUMNS: {
-  [key: string]: ColumnDef<ServiceFieldsFragment>;
-} = {
-  serviceDetails: {
-    header: 'Service Details',
-    render: (service) => (
-      <Stack>
-        {serviceDetails(service).map((s, i) => (
-          <Typography
-            // eslint-disable-next-line react/no-array-index-key
-            key={i}
-            variant='body2'
-            sx={{
-              whiteSpace: 'pre-wrap',
-              display: '-webkit-box',
-              WebkitBoxOrient: 'vertical',
-              WebkitLineClamp: '1',
-              overflow: 'hidden',
-              maxWidth: '300px',
-            }}
-          >
-            {s}
-          </Typography>
-        ))}
-      </Stack>
-    ),
-  },
-};
 
 const EnrollmentServicesPage = () => {
   const { enrollment } = useEnrollmentDashboardContext();
