+++ src/modules/services/components/ProjectServicesTable.tsx
@@ -3,9 +3,10 @@ import { useCallback, useMemo } from 'react';
 import { ColumnDef } from '@/components/elements/table/types';
 import ClientName from '@/modules/client/components/ClientName';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
-import { SERVICE_BASIC_COLUMNS } from '@/modules/enrollment/components/pages/EnrollmentServicesPage';
+
 import EnrollmentDateRangeWithStatus from '@/modules/hmis/components/EnrollmentDateRangeWithStatus';
 import { useFilters } from '@/modules/hmis/filterUtil';
+import { SERVICE_BASIC_COLUMNS } from '@/modules/services/serviceColumns';
 import { EnrollmentDashboardRoutes } from '@/routes/routes';
 import {
   GetProjectServicesDocument,
